reworked module structure and fixed wrong left over core imports
