--- Physics/ExpertSystem/expertsystem/state/validation.py
@@ -1,15 +1,0 @@
-import logging
-
-from core.topology.graph import (get_edges_ingoing_to_node,
-                                 get_edges_outgoing_to_node)
-
-from core.state.particle import (get_xml_label, XMLLabelConstants,
-                                 StateQuantumNumberNames,
-                                 InteractionQuantumNumberNames,
-                                 ParticlePropertyNames,
-                                 QNNameClassMapping,
-                                 QNClassConverterMapping)
-
-from core.state.propagation import (AbstractPropagator)
-
-
