+++ Physics/ExpertSystem/tests/unittests/conservation-laws/test_cparity.py
@@ -1,10 +1,10 @@
 from itertools import (product)
 
-from core.state.conservationrules import CParityConservation
-from core.state.particle import (StateQuantumNumberNames,
-                                 ParticlePropertyNames,
-                                 InteractionQuantumNumberNames,
-                                 Spin)
+from expertsystem.state.conservationrules import CParityConservation
+from expertsystem.state.particle import (StateQuantumNumberNames,
+                                         ParticlePropertyNames,
+                                         InteractionQuantumNumberNames,
+                                         Spin)
 
 
 class TestCParity(object):
